pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ROSE"
version = "0.1.0"
description = "Toolkit to express and execute ML surrogate building workflows on HPC"

authors = [
    {name = "Aymen Alsaadi", email = "aymen.alsaadi@rutgers.edu"},
    {name = "RADICAL Lab"},
]
maintainers = [
    {name = "Aymen Alsaadi", email = "aymen.alsaadi@rutgers.edu"},
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = ["numpy", "radical.asyncflow", "rhapsody-py"]

[project.urls]
Homepage = "https://github.com/radical-cybertools/ROSE"
Issues = "https://github.com/radical-cybertools/ROSE/issues"
Documentation = "https://radical-cybertools.github.io/ROSE/"

[project.optional-dependencies]
lint = ["ruff"]

# All test deps
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pre-commit",
]

doc = [
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.9",
    "mkdocstrings[python]>=0.26.1",
    "mkdocs-gen-files==0.5.0",
    "mkdocs-literate-nav==0.6.1",
    "mkdocs-section-index==0.3.8",
    "mkdocs-glightbox",
    "mkdocs-material-extensions",
    "mkdocs-minify-plugin"
]

[tool.ruff]
line-length = 100
target-version = "py39"
fix = true
exclude = ["examples/use_cases/"]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "N", "UP"]
# UP007/UP045: allow Optional[X] and X | Y syntax for py3.9 compat
# N803/N806: uppercase variable/arg names are standard ML convention (X, X_labeled, etc.)
# N801: allow non-CapWords class names for ML acronyms (MC_Dropout_CNN, etc.)
# N812/N813/N816/N817: allow ML import aliases (functional as F, etc.)
ignore = ["UP007", "UP045", "N803", "N806", "N801", "N812", "N813", "N816", "N817"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "6.0"

testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["rose"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.codespell]
skip = """
.git,
.github,
__pycache__,
build,
dist,
.*egg-info
"""
